Allow clicking mark in editor to show full error
Background compilations do not update the stderr display at the bottom
of the screen, because the user may not be interacting with the module
that has compiled. This commit allows the user to click the error marker
displayed in the left margin to display the stderr output, including
links and suggestions.